Refactor SkillQueue to treat row 1 as a component
